Add files via upload
Software the reads through HDDs with files from various drone types. The data is extracted from the file metadata. The data is then formatted before being  split into 4 four dataframes.  This will later become 4 maps in ArcGIS that are updated with each run. 

The script also prepare the information to create and update the filestructur within a seperate network. Further the scripts makes a log of the previous filestructures, fileupdates and file conversions that is used in the update process. 

This script provides in the input to two additional scripts "CopyPath.py" & "DroneMap" that complete the process inside an external network.