Add files via upload
Download data (pdf reports and txt) from the SMP database. The data is divided by municipality
and year. This script is connected to a second script that data mines and summarize the 
data.  

Additional development:
Create easier interface to allow selection of company categories/responsible organization  
Develop as a part of a GUI?
Further error search